backend/requirements.txt
asgiref==3.7.2
Django==4.2.1
sqlparse==0.4.4
typing_extensions==4.6.2

serializers~=0.2.4
djangorestframework~=3.14.0